change u128 to u64 for oracle calls, change save function

+++ tsp_samplers/src/algorithms.rs
@@ -1,6 +1,6 @@
 use rustc_hash::FxHashMap;
 
-pub type HillclimbFunction = fn(&Vec<u32>, &Vec<Vec<i32>>) -> (Vec<u32>, i32, u128);
+pub type HillclimbFunction = fn(&Vec<u32>, &Vec<Vec<i32>>) -> (Vec<u32>, i32, u64);
 pub type NodeMap = FxHashMap<Vec<u32>, (u32, i32)>;
 pub type EdgeMap = FxHashMap<(u32, u32), i32>;
 
@@ -17,5 +17,5 @@ pub use two_opt::{two_opt_besti, two_opt_firsti};
 
 pub trait SamplingAlg {
     fn get_hc_calls(&self) -> u64;
-    fn get_oracle_calls(&self) -> u128;
+    fn get_oracle_calls(&self) -> u64;
 }
